more or less finished chapter 12 with more to do. TODO updated with challenges and notes from end of chapter 12.

+++ src/HashMapStack.c
@@ -21,7 +21,7 @@ void* init_HashMapStack(void* inStk,void * arguments){
 			for(i=0;i<size;i++){
 				((HashMap**)stk->super.values)[i] = NULL;
 			}
-    stk->super.size = size;
+    stk->super.size = (int)size;
 	stk->super.top = &topHashMapStack;
 	stk->super.pop = &popHashMapStack;
 	stk->super.push = &pushHashMapStack;
